API/MainApp/requirements.txt
flask
flask-cors
python-dotenv
openai
mysql-connector-python